[issue #12] Improve robustness of populate-copied-stack-state.sh
- Check if directories exist before trying to delete them
  Alternative would have been to use --force, but that looks a bit more
error prone and less explicit when it is needed.
- Check if dcos package repository is already present before adding as 
  duplicate causes dcos command to fail.

+++ 04_implementation/scripts/populate-copied-stack-state.sh
@@ -5,8 +5,14 @@ cd `dirname ${BASH_SOURCE[0]}`
 
 if [ -e ${STATE_TAR_FILE} ]
 	then 
-		run rm -r ${HOME}/.dcos/
-		run rm -r ${CURRENT_STATE_DIR}
+		if [ -e ${HOME}/.dcos/ ]
+			then
+				run rm -r ${HOME}/.dcos/
+		fi
+		if [ -e ${CURRENT_STATE_DIR} ]
+			then
+				run rm -r ${CURRENT_STATE_DIR}
+		fi
 		echo "Populating state..."
 		run tar xzvf ${STATE_TAR_FILE}
